ragbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
elastic-transport==8.13.1
elasticsearch==8.13.1
numpy==1.26.4
tiktoken==0.7.0
orjson==3.10.3
ujson==5.9.0
starlette==0.37.2
typing-extensions==4.12.2

